+++ hm.js
@@ -54,6 +54,7 @@ Hooks.once('ready', async() => {
 
 Hooks.on('createActor', HMActor.createActor);
 Hooks.on('createToken', HMActor.createToken);
+Hooks.on('createItem', HMItem.createItem);
 Hooks.on('renderCombatTracker', HMCombatTracker.renderCombatTracker);
 Hooks.on('hotbarDrop', HMMacro.hotbarDrop);
 Hooks.on('diceSoNiceRollStart', (_messageId, context) => {