+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Fixed
+- Beasts no longer break if they have a language object.
+-
 ## [0.2.2] - 2022-04-15
 ### Added
 - Skill macros.