+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Fixed
 - Beasts no longer break if they have a language object.
--
+- Beasts receive bonus spellpoints as they should.
+
 ## [0.2.2] - 2022-04-15
 ### Added
 - Skill macros.