+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Alert when new wound damage exceeds ToP.
+### Changed
+- Defense results on a nat 1 were misleading.
 ### Fixed
 - Beasts no longer break if they have a language object.
 - Beasts receive bonus spellpoints as they should.